last commit

+++ users.php
@@ -9,10 +9,21 @@
     }
 
     $user_list = '';
+    $search = '';
 
     //getting the list of numbers
 
-    $query = "SELECT * FROM user WHERE is_deleted=0 ORDER BY first_name";
+    if(isset($_GET['search'])) {
+    
+        $search = mysqli_real_escape_string($connection, $_GET['search']);
+        $query = "SELECT * FROM user WHERE (first_name LIKE '%$search%' OR last_name LIKE '%$search%' OR email LIKE '%$search%') AND is_deleted=0 ORDER BY first_name";        
+    
+    } else {
+
+        $query = "SELECT * FROM user WHERE is_deleted=0 ORDER BY first_name";
+
+    }
+
     $users = mysqli_query($connection, $query);
 
     verify_query($users);
@@ -23,7 +34,7 @@
             $user_list .= "<td>{$user['last_name']}</td>";
             $user_list .= "<td>{$user['last_login']}</td>";
             $user_list .= "<td><a href=\"modify-user.php?user_id={$user['id']}\">Edit</a></td>";
-            $user_list .= "<td><a href=\"delete-user.php?user_id={$user['id']}\">Delete</a></td>";
+            $user_list .= "<td><a href=\"delete-user.php?user_id={$user['id']}\"onclick=\"return confirm('Are you sure?');\">Delete</a></td>";
             $user_list .= "</tr>";
             
             }
@@ -42,7 +53,15 @@
         <div class="loggedin">Welcome Username <?php echo $_SESSION['first_name']; ?><a href="logout.php">Log out</a></div>
     </header>
     <main>
-        <h1>Users<span><a href="add-user.php">+ Add New</a></span></h1>
+        <h1>Users<span><a href="add-user.php">+ Add New</a> | <a href="users.php">Refresh</a></span></h1>
+
+        <div class="search">
+            <form action="users.php" method="get">
+                <p>
+                <input type="text" name="search" id="" placeholder="Type First Name, Last Name or Email adderss and Press Enter" value="<?php echo $search; ?>" required autofocus>
+                </p>
+            </form>
+        </div>
         
         <table class="masterlist">
             <tr>
